+++ browser-extension/app/components/app/pastbl-main-button.vue
