+++ browser-extension/app/components/app/tabs/pastbl-create-tab.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="p-2">
+    <form @submit.prevent>
+      <textarea
+        id="asd"
+        name="asd"
+        class="textarea textarea-bordered w-full"
+      />
+    </form>
+  </div>
+</template>
