--- browser-extension/app/components/pastbl-create-tab.vue
@@ -1,11 +1,0 @@
-<template>
-  <div class="p-2">
-    <form @submit.prevent>
-        <textarea name="asd" id="asd" class="textarea textarea-bordered w-full">
-
-
-
-        </textarea>
-    </form>
-  </div>
-</template>
