+++ browser-extension/app/components/settings/pastbl-settings-button.vue
