+++ browser-extension/app/components/settings/pastbl-settings-color-form-control.vue
