+++ browser-extension/app/components/settings/pastbl-with-settings-tooltip.vue
